% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cstr}
\alias{cstr}
\title{Continuous stirred tank reactor data (idframe)}
\format{an \code{idframe} object with 7500 samples, one input and two
outputs}
\usage{
cstr
}
\description{
The Process is a model of a Continuous Stirring Tank Reactor, 
where the reaction is exothermic and the concentration is 
controlled by regulating the coolant flow.
\cr
}
\details{
Inputs: q, Coolant Flow l/min
Outputs:
\describe{
\item{Ca}{Concentration mol/l}
\item{T}{Temperature Kelvin}}
}
\keyword{datasets}

