% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cstr_mis}
\alias{cstr_mis}
\title{Continuous stirred tank reactor data with missing values}
\format{an \code{idframe} object with 7500 samples, one input and two
outputs}
\usage{
cstr_mis
}
\description{
This dataset is derived from the \code{cstr} dataset with few samples
containing missing values, in one or all variables. It is used to 
demonstrate the capabilities of the \code{misdata} routine.
}
\seealso{
\code{\link{cstr}}, \code{\link{misdata}}
}
\keyword{datasets}

