% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cstrData}
\alias{cstrData}
\title{Continuous stirred tank reactor data (data.frame)}
\format{an \code{data.frame} object with 7500 rows and three columns:
q, Ca and T}
\source{
\url{ftp://ftp.esat.kuleuven.be/pub/SISTA/data/process_industry/cstr.dat.gz}
}
\usage{
cstrData
}
\description{
The Process is a model of a Continuous Stirring Tank Reactor, 
where the reaction is exothermic and the concentration is 
controlled by regulating the coolant flow.
\cr
}
\details{
Inputs: q, Coolant Flow l/min
Outputs:
\describe{
\item{Ca}{Concentration mol/l}
\item{T}{Temperature Kelvin}}
}
\keyword{datasets}

