% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/preProcess.R
\name{dataImpute}
\alias{dataImpute}
\title{Impute missing values}
\usage{
dataImpute(data, m = 5)
}
\arguments{
\item{data}{an object of class \code{idframe}}

\item{m}{the number of imputed datasets to create.}
}
\value{
A list containing the following elements

\tabular{ll}{
   \code{imputations} \tab a list of length \code{m} with an imputed dataset in each
   entry. Each element is of class \code{idframe} \cr
   \code{m} \tab an integer indicating the number of imputations run
 }
}
\description{
Uses the bootstrap EM Algorithm on incomplete data and creates imputed datasets
}
\details{
This function uses the \code{\link[Amelia]{amelia}} function, provided by the
\pkg{amelia} package, to perform multiple imputation.
}
\references{
Honaker, J., King, G., Blackwell, M. (2011). Amelia II: A Program for Missing Data.
Journal of Statistical Software, 45(7), 1–47. URL http://www.jstatsoft.org/v45/i07/.
}
\seealso{
\code{\link[Amelia]{amelia}}
}

