% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/partition.R
\name{dataPartition}
\alias{dataPartition}
\title{Split data into training and validation sets}
\usage{
dataPartition(object, p = 0.6)
}
\arguments{
\item{object}{an object of class \code{idframe}}

\item{p}{the percentage of the data that goes to training (Default : \code{0.6})}
}
\value{
list containing estimation and validation idframe objects
}
\description{
The function splits the data into training and validation sets and returns them bundled
as a list. The size of the sets are determined by the parameter \code{p}.
}
\examples{
data(cstr)
splitList <- dataPartition(cstr,p=0.6)
train <- splitList$estimation # training set
test <- splitList$validation # testing set
}

