% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/demean.R
\name{demean}
\alias{demean}
\title{Mean-Center the data}
\usage{
demean(data)
}
\arguments{
\item{data}{an object of class \code{idframe}}
}
\value{
A list containing the following elements

\tabular{ll}{
   \code{fitted.values} \tab \code{idframe} object with mean-centered variables \cr
   \code{output.mean} \tab \code{vector} containing means for each output variable \cr
   \code{input.mean} \tab \code{vector} containing means for each input variable
 }
}
\description{
Mean Centers the input and output matrices.
}
\examples{
data(cstr)
fit.mean <- demean(cstr)
cstr_demean <- predict(fit.mean)
}
\seealso{
\code{\link{predict.demean}}, \code{\link[stats]{colMeans}}
}

