% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/detrend.R
\name{detrend.idframe}
\alias{detrend.idframe}
\title{Remove linear trends}
\usage{
detrend.idframe(data)
}
\arguments{
\item{data}{an object of class \code{idframe}}
}
\value{
A list containing the following elements

\tabular{ll}{
   \code{fitted.values} \tab \code{idframe} object with detrended variables \cr
   \code{output.trend} \tab \code{list} containing trend fits for each output variable \cr
   \code{input.trend} \tab \code{list} containing trend fits for each input variable
 }
}
\description{
Removes the linear function from the input and output matrices.
}
\examples{
data(cstr)
fit <- detrend.idframe(cstr)
cstr_detrend <- predict(fit)
}
\seealso{
\code{\link{predict.detrend.idframe}}, \code{\link[stats]{lm}}
}

