% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preprocess.R
\name{detrend}
\alias{detrend}
\title{Remove offsets and linear trends}
\usage{
detrend(x, type = c("constant", "linear")[1])
}
\arguments{
\item{x}{an object of class \code{idframe}}

\item{type}{trend type - "constant" or "linear". (Default: \code{"constant"})}
}
\value{
A list containing the following elements

\tabular{ll}{
   \code{fitted.values} \tab \code{idframe} object with detrended variables \cr
   \code{output_trend} \tab \code{list} containing trend fits for each output
   variable \cr
   \code{input_trend} \tab \code{list} containing trend fits for each input
   variable
 }
}
\description{
Removes the offsets or linear trends in each of the input and output matrices.
}
\examples{
data(cstr)
fit <- detrend(cstr,type="linear") # remove linear trends
Zdetrend <- predict(fit) # get the detrended data

demean <- detrend(cstr) # remove offsets
Zcent <- predict(demean) # get the centered data
}
\seealso{
\code{\link{predict.detrend}}, \code{\link[stats]{lm}}
}

