% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estpoly.R
\name{estpoly}
\alias{estpoly}
\title{Estimated polynomial object}
\usage{
estpoly(sys, fitted.values, residuals, options = NULL, call, stats,
  termination = NULL, input)
}
\arguments{
\item{sys}{an \code{idpoly} object containing the estimated polynomial
coefficients}

\item{fitted.values}{1-step ahead predictions on the training dataset}

\item{residuals}{1-step ahead prediction errors}

\item{options}{optimization specification ser used (applicable for non-linear least
squares)}

\item{call}{the matched call}

\item{stats}{a list containing estimation statistics}

\item{termination}{termination criteria for optimization}

\item{input}{input signal of the training data-set}
}
\description{
Estimated discrete-time polynomial model returned from an estimation
routine.
}
\details{
Do not use \code{estpoly} for directly specifing an input-output polynomial model.
\code{\link{idpoly}} is to be used instead
}

