% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estpoly.R
\name{fitch}
\alias{fitch}
\title{Fit Characteristics}
\usage{
fitch(x)
}
\arguments{
\item{x}{the estimated model}
}
\value{
A list containing the following elements

\item{MSE}{Mean Square Error measure of how well the response of the model fits
the estimation data}
\item{FPE}{Final Prediction Error}
\item{FitPer}{Normalized root mean squared error (NRMSE) measure of how well the 
response of the model fits the estimation data, expressed as a percentage.}
\item{AIC}{Raw Akaike Information Citeria (AIC) measure of model quality}
\item{AICc}{Small sample-size corrected AIC}
\item{nAIC}{Normalized AIC}
\item{BIC}{Bayesian Information Criteria (BIC)}
}
\description{
Returns quantitative assessment of the estimated model as a list
}

