% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{frd}
\alias{frd}
\title{Frequency response data}
\format{an \code{idfrd} object with response at 128 frequency points}
\usage{
frd
}
\description{
This dataset contains frequency response data of an unknown SISO system.
}
\keyword{datasets}

