% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estUtil.R
\name{getcov}
\alias{getcov}
\title{Parameter covariance of the identified model}
\usage{
getcov(sys)
}
\arguments{
\item{sys}{a linear, identified parametric model}
}
\description{
Obtain the parameter covariance matrix of the linear, identified 
parametric model
}

