% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{\%=\%}
\alias{\%=\%}
\alias{g}
\title{Multiple assignment operator}
\usage{
l \%=\% r
}
\arguments{
\item{l}{the variables to be assigned}

\item{r}{the list or function-return object}

\item{\ldots}{other arguments}
}
\description{
Assign multiple variables from a list or function return object
}

