% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/idframe.R
\name{idframe}
\alias{idframe}
\title{S3 class for storing input-output data.}
\usage{
idframe(output = numeric(0), input = numeric(0), type = c("time",
  "freq")[1], Ts = 1, t.start = 0, t.end = NA, tUnit = "seconds",
  frequencies = NA, fUnit = "Hz")
}
\arguments{
\item{output}{dataframe/matrix/vector containing the outputs}

\item{input}{dataframe/matrix/vector containing the inputs}

\item{type}{indicates the domain of the data (Default:"time")}

\item{Ts}{sampling interval (Default: 1)}

\item{t.start}{Starting time (Valid only if type="time")}

\item{t.end}{End time. Optional Argument (Valid only if type="time")}

\item{tUnit}{Time Unit (Default: "seconds")}

\item{frequencies}{Vector containing the list of frequencies at which the data was
recorded (Valid only if type="frequency")}

\item{fUnit}{Frequency Unit (Valid only if type="frequency")}
}
\value{
an idframe object
}
\description{
\code{idframe} is an S3 class for storing and manipulating input-ouput data. It supports discrete time and frequency domain data.
}
\examples{
dataMatrix <- matrix(rnorm(1000),ncol=5)
data <- idframe(output=dataMatrix[,3:5],input=dataMatrix[,1:2],Ts=1)
}
\seealso{
\code{\link{plot.idframe}}, the plot method for idframe objects,
\code{\link{summary.idframe}}, the summary method for idrame objects
}

