% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/impulse.R
\name{impulse}
\alias{impulse}
\title{Impulse Response Plots}
\usage{
impulse(model)
}
\arguments{
\item{model}{an object of class \code{impulseest}}
}
\description{
Plots the estimated IR Coefficients
}
\seealso{
\code{\link{impulseest}},\code{\link{step}}
}

