% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/impulse.R
\name{impulseest}
\alias{impulseest}
\title{Estimate Impulse Response Models}
\usage{
impulseest(data, M = 30, K)
}
\arguments{
\item{data}{an object of class \code{idframe}}

\item{M}{Order of the FIR Model (Default:\code{30})}

\item{K}{Transport delay in the estimated impulse response}
}
\description{
\code{impulseest} is used to estimate impulse response models in the
given data
}
\seealso{
\code{\link{plot.impulseest}}, \code{\link{step}}
}

