% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ioNamesData.R
\name{inputData}
\alias{inputData}
\alias{inputData.idframe}
\alias{outputData}
\alias{outputData.idframe}
\title{Output or Input-data}
\usage{
inputData(x, series)
}
\arguments{
\item{x}{\code{idframe} object}

\item{series}{the indices to extract}
}
\description{
Extract output-data or input-data in idframe objects
}

