% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ioNamesData.R
\name{inputNames<-}
\alias{inputNames}
\alias{inputNames<-}
\alias{inputNames<-.idframe}
\alias{outputNames}
\alias{outputNames<-}
\alias{outputNames<-.idframe}
\title{Extract or set series' names}
\usage{
inputNames(x) <- value
}
\arguments{
\item{x}{\code{idframe} object}

\item{value}{vector of strings}
}
\description{
Extract or set names of series in input or output
}

