% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ioNamesData.R
\name{nInputSeries}
\alias{nInputSeries}
\alias{nOutputSeries}
\title{Number of series in input or output}
\usage{
nInputSeries(data)
}
\arguments{
\item{data}{\code{idframe} object}
}
\description{
Number of series in input or output in a idframe object
}

