% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/idframe.R
\name{plot.idframe}
\alias{plot.idframe}
\title{Plotting idframe objects}
\usage{
\method{plot}{idframe}(object, par = list(mar = c(3, 4, 2, 2)),
  col = "steelblue", ...)
}
\arguments{
\item{object}{an object of class \code{idframe}}

\item{par}{a list of arguments passed to par() before plotting.}

\item{col}{line color, to be passed to plot.(Default=\code{"steelblue"})}

\item{...}{additional arguments to be passed to the \code{tfplot} function}
}
\description{
Plotting method for objects inherting from class \code{idframe}
}
\examples{
data(cstr)
plot(cstr,col="blue")
}
\seealso{
\code{\link[tfplot]{tfplot}}
}

