% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idframe.R
\name{plot.idfrd}
\alias{plot.idfrd}
\title{Plotting idfrd objects}
\usage{
\method{plot}{idfrd}(x, col = "steelblue", lwd = 1)
}
\arguments{
\item{x}{An object of class \code{idframe}}

\item{col}{a specification for the line colour (Default : \code{"
steelblue"})}

\item{lwd}{the line width, a positive number, defaulting to 1}
}
\description{
Generates the bode plot of the given frequency response data. It uses the
ggplot2 plotting engine
}
\examples{
data(frd)
plot(frd)

}
\seealso{
\code{\link[ggplot2]{ggplot}}
}

