% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparam.R
\name{plot.impulseest}
\alias{plot.impulseest}
\title{Impulse Response Plots}
\usage{
\method{plot}{impulseest}(model, sig = 0.975)
}
\arguments{
\item{model}{an object of class \code{impulseest}}

\item{sig}{Significance Limits (Default: \code{0.975})}
}
\description{
Plots the estimated IR coefficients along with the significance limits
at each lag.
}
\seealso{
\code{\link{impulseest}},\code{\link{step}}
}

