% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/demean.R
\name{predict.demean}
\alias{predict.demean}
\title{Predict the centered values}
\usage{
\method{predict}{demean}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{idframe}}

\item{newdata}{An optional idframe object in which to look for variables with which
to predict. If ommited, the original idframe object is used}
}
\value{
an \code{idframe} object
}
\description{
Center an \code{idframe} object based on the training center means
}
\examples{
## Examples for train and test sets
data(cstr)
splitList <- dataPartition(cstr,p=0.6)
train <- splitList$estimation # training set
test <- splitList$validation # testing set
fit.mean <- demean(train)
train_demean <- predict(fit.mean)
test_demean <- predict(fit,newdata=test)
}

