% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/detrend.R
\name{predict.detrend.idframe}
\alias{predict.detrend.idframe}
\title{Predict method for trend fits on idframe objects}
\usage{
\method{predict}{detrend.idframe}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{idframe}}

\item{newdata}{An optional idframe object in whic to look for variables with which
to predict. If ommited, the original detrended idframe object is used}
}
\value{
an \code{idframe} object
}
\description{
Detrended \code{idframe} object based on linear trend fit
}
\examples{
## Examples for train and test sets
data(cstr)
splitList <- dataPartition(cstr,p=0.6)
train <- splitList$estimation # training set
test <- splitList$validation # testing set
fit <- detrend.idframe(train)
train_detrend <- predict(fit)
test_detrend <- predict(fit,newdata=test)
}

