% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preprocess.R
\name{predict.detrend}
\alias{predict.detrend}
\title{Detrend data based on linear trend fits}
\usage{
\method{predict}{detrend}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{idframe}}

\item{newdata}{An optional idframe object in which to look for variables with
which to predict. If ommited, the original detrended idframe object is used}
}
\value{
an \code{idframe} object
}
\description{
Returns detrended \code{idframe} object based on linear trend fit
}
\examples{
data(cstr)
train <- dataSlice(cstr,end=5000)
test <- dataSlice(cstr,start=6001)
fit <- detrend(train)
Ztrain <- predict(fit)
Ztest <- predict(fit,test)
}

