% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/tf.R
\name{print.tf}
\alias{print.tf}
\title{Display the Transfer Function}
\usage{
\method{print}{tf}(G)
}
\arguments{
\item{G}{an object of class \code{tf}}
}
\description{
Printing method for objects inheriting from class \code{tf}
}

