% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readData.R
\name{read.idframe}
\alias{read.idframe}
\title{Data input into a idframe object}
\usage{
read.idframe(data, ninputs = 1, type = c("time", "freq")[1], Ts = 1,
  freqData = FALSE)
}
\arguments{
\item{data}{a \code{data.frame} object}

\item{ninputs}{the number of input columns. (Default: 1)}

\item{type}{indicates the domain of the data (Default:\code{"time"})}

\item{Ts}{sampling interval (Default: 1)}

\item{freqData}{a logical value indicating whether the file contains the list of
frequencies. If \code{TRUE}, they need to be present in the first column.
(Default: \code{idframe})}
}
\value{
an idframe object
}
\description{
Read the contents of an file in table format into a \code{idframe} object.
}
\details{
If \code{type="freq"} and \code{freqData = TRUE}, then the first column in the file
should contain the frequencies.
}

