% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{read.odf.idframe}
\alias{read.odf.idframe}
\title{Reading from .odf documents}
\usage{
read.odf.idframe(file, sheetName, header = TRUE, ninputs = 0, Ts = 1,
  unit = c("seconds", "minutes", "hours", "days")[1], ...)
}
\arguments{
\item{file}{the path to the file to read}

\item{sheetName}{a character string with the sheet name}

\item{header}{a logical value indicating whether the first row corresponding to 
the first element of the rowIndex vector contains the names of the variables.}

\item{ninputs}{the number of input columns. (Default: 0)}

\item{Ts}{sampling interval (Default: 1)}

\item{unit}{Time Unit (Default: "seconds")}

\item{...}{additional arguments to be passed to the \code{\link[xlsx]{read.xlsx2}} function}
}
\value{
an idframe object
}
\description{
Read the contents of an a .odf document into a \code{idframe} object.
}
\details{
The \code{read.odf.idframe} function uses the \code{\link[gnumeric]{read.gnumeric.sheet}} function, 
provided by the \pkg{xlsx} package, to read data from a .odf file and then calls the 
\code{\link{read.idframe}} function to read the data into a idframe object
}
\seealso{
\code{\link[xlsx]{read.xlsx2}}
}

