% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readData.R
\name{readxls.idframe}
\alias{readxls.idframe}
\title{Read the contents of a worksheet into a idframe object}
\usage{
readxls.idframe(file, sheetName, header = TRUE, ninputs = 1,
  type = c("time", "freq")[1], Ts = 1, freqData = FALSE, ...)
}
\arguments{
\item{file}{the path to the file to read}

\item{sheetName}{a character string with the sheet name}

\item{header}{a logical value indicating whether the first row corresponding to
the first element of the rowIndex vector contains the names of the variables.
(Default: \code{TRUE})}

\item{ninputs}{the number of input columns. (Default: 1)}

\item{type}{ndicates the domain of the data (Default:\code{"time"})}

\item{Ts}{sampling interval (Default: 1)}

\item{freqData}{a logical value indicating whether the file contains the list of
frequencies. If \code{TRUE}, they need to be present in the first column.
(Default: \code{idframe})}

\item{...}{additional arguments to be passed to the \code{\link[xlsx]{read.xlsx2}} function}
}
\value{
an idframe object
}
\description{
Read the contents of an excel worksheet into a \code{idframe} object.
}
\details{
The \code{readxlsx.idframe} function uses the \code{\link[xlsx]{read.xlsx2}} function,
provided by the \pkg{xlsx} package, to read data from an excel file and then calls the
\code{\link{read.idframe}} function to read the data into a idframe object


If \code{type="freq"} and \code{freqData = TRUE}, then the first column in the file
should contain the frequencies.


The function requires the java runtime to be installed on the system (Requirement of
the \pkg{xlsx} package).
}
\seealso{
\code{\link[xlsx]{read.xlsx2}}
}

