% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estpoly.R
\name{residplot}
\alias{residplot}
\title{Plot residual characteristics}
\usage{
residplot(model, newdata = NULL)
}
\arguments{
\item{model}{estimated polynomial model}

\item{newdata}{an optional dataset on which predictions are to be computed. If
not supplied, predictions are computed on the training dataset.}
}
\description{
Computes the 1-step ahead prediction errors (residuals) for an estimated polynomial
model, and plots auto-correlation of the residuals and the 
cross-correlation of the residuals with the input signals.
}

