% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim.R
\name{sim.arx}
\alias{sim.arx}
\title{Simulate from an ARX Model}
\usage{
\method{sim}{arx}(model, input, sigma = 0, seed = NULL)
}
\arguments{
\item{model}{an object of class \code{arx} containing the coefficients}

\item{input}{a vector/matrix containing the input}

\item{sigma}{standard deviation of the innovations (Default= \code{0})}

\item{seed}{integer indicating the seed value of the random number generator}
}
\value{
a vector containing the output
}
\description{
Simulate the response of an ARX system, given the input
}
\details{
The routine is currently built only for SISO systems. Future Versions will
include support for MIMO systems
}
\examples{
u <- rnorm(200,sd=1)
model <- arx(A=c(1,-1.5,0.7),B=c(0.8,-0.25),ioDelay=1)
y <- sim(model,u,sigma=0.1)
}
\seealso{
\code{\link{arx}} for defining ARX models
}

