% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sim.R
\name{sim}
\alias{sim}
\title{Simulate dynamic system}
\usage{
sim(model, input, sigma = 0, seed = NULL)
}
\arguments{
\item{model}{the system model to simulate}

\item{input}{a vector/matrix containing the input}

\item{sigma}{standard deviation of the innovations (Default= \code{0})}

\item{seed}{integer indicating the seed value of the random number generator}
}
\value{
a vector containing the output
}
\description{
Simulate the response of a system given the input
}
\details{
The routine is currently built only for SISO systems. Future Versions will
include support for MIMO systems. Current support
}
\seealso{
\code{\link{sim.idpoly}} for simulating polynomial models
}

