% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparam.R
\name{step}
\alias{step}
\title{Step Response Plots}
\usage{
step(model)
}
\arguments{
\item{model}{an object of class \code{impulseest}}
}
\description{
Plots the step response of a system, given the IR model
}
\examples{
uk <- rnorm(1000,1)
yk <- filter (uk,c(0.9,-0.4),method="recursive") + rnorm(1000,1)
data <- idframe(output=data.frame(yk),input=data.frame(uk))
fit <- impulseest(data)
step(fit) 

}
\seealso{
\code{\link{impulseest}}
}

