% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/idframe.R
\name{summary.idframe}
\alias{summary.idframe}
\title{idframe-object summaries}
\usage{
\method{summary}{idframe}(object)
}
\arguments{
\item{object}{an object of class \code{idframe}}
}
\description{
Generates a summary of objects inherting from class \code{idframe}
}
\examples{
data(cstr)
summary(cstr)
}

