% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tf.R
\name{tf}
\alias{tf}
\title{S3 class for transfer functions}
\usage{
tf(num = c(1), den = c(1), Ts = 1)
}
\arguments{
\item{num}{coefficients of the numerator plynomial in q^{-1}}

\item{den}{coefficients of the denominator plynomial in q^{-1}}

\item{Ts}{sampling time}
}
\value{
an object of class tf
}
\description{
\code{tf} is an S3 class for defining transfer functions.
}

