% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idframe.R
\name{time}
\alias{deltat}
\alias{frequency}
\alias{time}
\title{Sampling times of IO data
 
\code{time} creates the vector of times at which data was sampled. \code{frequency} returns the number of damples per unit time and \code{deltat} the time-interval 
between observations}
\usage{
time(x)
}
\arguments{
\item{x}{a idframe object, or a univariate or multivariate time-series, or a vector or matrix}
}
\description{
Sampling times of IO data
 
\code{time} creates the vector of times at which data was sampled. \code{frequency} returns the number of damples per unit time and \code{deltat} the time-interval 
between observations
}

