% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/partition.R
\name{dataSlice}
\alias{dataSlice}
\title{Subset an idframe data}
\usage{
dataSlice(data, indices)
}
\arguments{
\item{data}{an object of class \code{idframe}}

\item{indices}{the indices that need to be subsetted}
}
\description{
Subsetting method for datas of class \code{idframe}
}

