% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{detrend}
\alias{detrend}
\title{Remove offsets and linear trends}
\usage{
detrend(x, type = c("constant", "linear")[1])
}
\arguments{
\item{x}{an object of class \code{idframe}}

\item{type}{trend type - "constant" or "linear". (Default: \code{"constant"})}
}
\value{
A list containing the following elements
\item{fitted.values}{\code{idframe} object with detrended variables}
\item{output_trend}{\code{list} containing trend fits for each output 
   variable}
\item{input_trend}{\code{list} containing trend fits for each input 
   variable}
}
\description{
Removes the offsets or linear trends in each of the input and output matrices.
}
\examples{
data(cstr)
fit <- detrend(cstr,type="linear") # remove linear trends 
Zdetrend <- predict(fit) # get the detrended data

demean <- detrend(cstr) # remove offsets
Zcent <- predict(demean) # get the centered data

}
\seealso{
\code{\link{predict.detrend}}, \code{\link[stats]{lm}}
}

