% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{\%=\%}
\alias{\%=\%}
\title{Multiple assignment operator}
\usage{
l \%=\% r
}
\arguments{
\item{lhs, rhs}{the variables to be assingned, and the list or 
function return object}
}
\description{
Assign multiple variables from a list or function return object
}

