% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/impulse.R
\name{impulseest}
\alias{impulseest}
\title{Estimate Impulse Response Models}
\usage{
impulseest(data, lags = 30, conf = 0.95)
}
\arguments{
\item{data}{an object of class \code{idframe}}

\item{lags}{The number of lags upto which the estimate is to be
calculated. (Default:\code{30})}

\item{conf}{The confidence interval}
}
\value{
An object of class \code{impulseest} containing the following elements
\tabular{ll}{
  \code{coefs} \tab \code{list} containing the impulse response coefficients \cr
  \code{lower} \tab \code{list} containing the lower confidence bounds of the
  impulse response coefficients \cr
  \code{upper} \tab \code{list} containing the upper confidence bounds of the
  impulse response coefficients \cr
}
}
\description{
\code{impulseest} is used to estimate impulse response models given data
}
\details{
This function extends the \code{\link[vars]{irf}} function, which estimates the
impulse response coefficients given a VAR model for the data. The VAR model is fit
using the \code{\link[vars]{VAR}} function.
}
\seealso{
\code{\link[vars]{irf}}, \code{\link[vars]{VAR}}, \code{\link{impulse}} ,
\code{\link{step}}
}

