% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/impulse.R
\name{impulseest}
\alias{impulseest}
\title{Estimate Impulse Response Models}
\usage{
impulseest(data, M = 30, K = 0, regul = F, lambda = 1)
}
\arguments{
\item{data}{an object of class \code{idframe}}

\item{M}{Order of the FIR Model (Default:\code{30})}

\item{K}{Transport delay in the estimated impulse response
(Default:\code{0})}

\item{regul}{Parameter indicating whether regularization should be
used. (Default:\code{FALSE})}

\item{lambda}{The value of the regularization parameter. Valid only if
\code{regul=TRUE}. (Default:\code{1})}
}
\description{
\code{impulseest} is used to estimate impulse response models in the
given data
}
\examples{
uk <- rnorm(1000,1)
yk <- filter (uk,c(0.9,-0.4),method="recursive") + rnorm(1000,1)
data <- idframe(output=data.frame(yk),input=data.frame(uk))
fit <- impulseest(data)
plot(fit)
}
\seealso{
\code{\link{step}}
}

