% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estUtil.R
\name{optimOptions}
\alias{optimOptions}
\title{Create optimization options}
\usage{
optimOptions(tol = 1e-05, maxIter = 20, LMinit = 2, LMstep = 2)
}
\arguments{
\item{tol}{Minimum ratio of the improvement to the current loss 
function. Iterations stop if this ratio goes below the tolerance
limit (Default: \code{1e-5})}

\item{maxIter}{Maximum number of iterations to be performed}

\item{LMinit}{Starting value of search-direction length 
in the Levenberg-Marquardt method.}

\item{LMstep}{Size of the Levenberg-Marquardt step}
}
\description{
Specify optimization options that are to be passed to the 
numerical estimation routines
}

