% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idframe.R
\name{plot.idframe}
\alias{plot.idframe}
\title{Plotting idframe objects}
\usage{
\method{plot}{idframe}(x, col = "steelblue", lwd = 1, main = NULL,
  size = 12, ...)
}
\arguments{
\item{x}{an \code{idframe} object}

\item{col}{line color, to be passed to plot.(Default=\code{"steelblue"})}

\item{lwd}{line width, in millimeters(Default=\code{1})}

\item{main}{the plot title. (Default = \code{NULL})}

\item{size}{text size (Default = \code{12})}

\item{\ldots}{additional arguments}
}
\description{
Plotting method for objects inherting from class \code{idframe}
}
\examples{
data(cstr)
plot(cstr,col="blue")

}

