% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/idframe.R
\name{plot.idframe}
\alias{plot.idframe}
\title{Plotting idframe objects}
\usage{
\method{plot}{idframe}(object, ...)
}
\arguments{
\item{object}{an object of class \code{idframe}}

\item{...}{additional arguments to be passed to the \code{plot} function}
}
\description{
Plotting method for objects inherting from class \code{idframe}
}
\examples{
data(distill)
plot(distill,col="blue")
}

