% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/idframe.R
\name{plot.idframe}
\alias{plot.idframe}
\title{Plotting idframe objects}
\usage{
\method{plot}{idframe}(object, ...)
}
\arguments{
\item{object}{an object of class "idframe"}

\item{...}{additional arguments to be passed to the `plot` function}
}
\description{
Plotting method for objects inherting from class "idframe"
}

