% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/idframe.R
\name{plot.idfrd}
\alias{plot.idfrd}
\title{Plotting idfrd objects}
\usage{
\method{plot}{idfrd}(object)
}
\arguments{
\item{object}{An object of class \code{idframe}}
}
\description{
Generates the bode plot of the given frequency response data. It uses the
ggplot2 plotting engine
}
\examples{
data(frf)
frf <- spa(data) # Estimates the frequency response from data
plot(frf)
}
\seealso{
\code{\link[ggplot2]{ggplot}}
}

