% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/idframe.R
\name{plot.idfrd}
\alias{plot.idfrd}
\title{Plotting idfrd objects}
\usage{
\method{plot}{idfrd}(object)
}
\arguments{
\item{object}{An object of class \code{idframe}}
}
\description{
Plot method for frequency resopnse data
}

