% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/impulse.R
\name{plot.impulseest}
\alias{plot.impulseest}
\title{Impulse Response Plots}
\usage{
\method{plot}{impulseest}(model)
}
\arguments{
\item{model}{an object of class \code{impulseest}}
}
\description{
Plots the estimated IR Coefficients
}
\seealso{
\code{\link{impulseest}},\code{\link{step}}
}

