% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/detrend.R
\name{predict.detrend}
\alias{predict.detrend}
\title{Predict method for trend fits on idframe objects}
\usage{
\method{predict}{detrend}(object, newdata = NULL, ...)
}
\arguments{
\item{object}{an object of class \code{idframe}}

\item{newdata}{An optional idframe object in which to look for variables with
which to predict. If ommited, the original detrended idframe object is used}
}
\value{
an \code{idframe} object
}
\description{
Detrended \code{idframe} object based on linear trend fit
}
\examples{
data(cstr)
train <- dataSlice(cstr,end=5000) # subset the first 5000 indices
test <- dataSlice(distill,start=6001) # subset from index 6001 till the end
fit <- detrend(train)
Ztrain <- predict(fit)
Ztest <- predict(fit,test)
}

