% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readData.R
\name{read.idframe}
\alias{read.idframe}
\title{Data input into a idframe object}
\usage{
read.idframe(data, ninputs = NULL, Ts = 1, unit = c("seconds", "minutes",
  "hours", "days")[1])
}
\arguments{
\item{data}{a \code{data.frame} object}

\item{ninputs}{the number of input columns. (Default: 0)}

\item{Ts}{sampling interval (Default: 1)}

\item{unit}{Time Unit (Default: "seconds")}
}
\value{
an idframe object
}
\description{
Read the contents of a data.frame/matrix into a \code{idframe} object.
}
\examples{
data(cstrData)
data <- read.idframe(cstrData,ninputs=1,Ts= 1,unit="minutes")

}

