% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/readData.R
\name{read.table.idframe}
\alias{read.table.idframe}
\title{Data input into a idframe object}
\usage{
read.table.idframe(file, header = TRUE, sep = ",", ninputs = 1,
  type = c("time", "freq")[1], Ts = 1, freqData = FALSE, tUnit = "time",
  ...)
}
\arguments{
\item{file}{the path to the file to read}

\item{header}{a logical value indicating whether the first row corresponding to
the first element of the rowIndex vector contains the names of the variables.
(Default: \code{TRUE})}

\item{sep}{the field separator character. Values on each line of the file are
separated by this character. (Default: \code{","})}

\item{ninputs}{the number of input columns. (Default: 1)}

\item{type}{indicates the domain of the data (Default:\code{"time"})}

\item{Ts}{sampling interval (Default: 1)}

\item{freqData}{a logical value indicating whether the file contains the list of
frequencies. If \code{TRUE}, they need to be present in the first column.
(Default: \code{idframe})}

\item{tUnit}{Time Unit (Default: "seconds")}

\item{...}{additional arguments to be passed to the \code{\link[utils]{read.table}} function}
}
\value{
an idframe object
}
\description{
Read the contents of an file in table format into a \code{idframe} object.
}
\details{
The \code{readxlsx.idframe} function uses the \code{\link[utils]{read.table}} function,
provided by the \pkg{xlsx} package, to read data from an excel file and then calls the
\code{\link{read.idframe}} function to read the data into a idframe object


If \code{type="freq"} and \code{freqData = TRUE}, then the first column in the file
should contain the frequencies.
}
\seealso{
\code{\link[utils]{read.table}}
}

