% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/readData.R
\name{read.xls.idframe}
\alias{read.xls.idframe}
\title{Read the contents of a worksheet into a idframe object}
\usage{
read.xls.idframe(file, sheetName, header = TRUE, ninputs = 0, Ts = 1,
  unit = c("seconds", "minutes", "hours", "days")[1], ...)
}
\arguments{
\item{file}{the path to the file to read}

\item{sheetName}{a character string with the sheet name}

\item{ninputs}{the number of input columns. (Default: 0)}

\item{Ts}{sampling interval (Default: 1)}

\item{unit}{Time Unit (Default: "seconds")}

\item{...}{additional arguments to be passed to the \code{\link[xlsx]{read.xlsx2}} function}
}
\value{
an idframe object
}
\description{
Read the contents of an excel worksheet into a \code{idframe} object.
}
\details{
The \code{read.xlsx.idframe} function uses the \code{\link[xlsx]{read.xlsx2}} function,
provided by the \pkg{xlsx} package, to read data from an excel file and then calls the
\code{\link{read.idframe}} function to read the data into a idframe object

The function requires the java runtime to be installed on the system (Requirement of
the \pkg{xlsx} package).
}
\examples{
library(xlsx)
dataMatrix <- data.frame(matrix(rnorm(1000),ncol=5))
colnames(dataMatrix) <- c("u1","u2","y1","y2","y3")
write.xlsx2(dataMatrix,file="test.xlsx",row.names=FALSE)

data <- read.xls.idframe("test.xlsx","Sheet1",ninputs=2,unit="minutes")
}
\seealso{
\code{\link[xlsx]{read.xlsx2}}
}

