% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim}
\alias{sim}
\title{Simulate dynamic system}
\usage{
sim(model, input, innov = NULL, sigma = 0, seed = NULL)
}
\arguments{
\item{model}{the linear system to simulate}

\item{input}{a vector/matrix containing the input}

\item{innov}{an optional times series of innovations. If not provided,
innovations are generated using the \code{rnorm} function}

\item{sigma}{standard deviation of the innovations (Default= \code{0})}

\item{seed}{integer indicating the seed value of the random number generator}
}
\value{
a vector containing the output
}
\description{
Simulate the response of a system given the input
}
\details{
The routine is currently built only for SISO systems. Future Versions will
include support for MIMO systems.
}
\examples{
# ARX Model
u <- rnorm(200,sd=1)
model <- idpoly(A=c(1,-1.5,0.7),B=c(0.8,-0.25),ioDelay=1)
y <- sim(model,u,sigma=0.1)

}

