% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/impulse.R
\name{step}
\alias{step}
\title{Step Response Plots}
\usage{
step(model)
}
\arguments{
\item{model}{an object of class \code{impulseest}}
}
\description{
Plots the step response of a system, given the IR model
}
\seealso{
\code{\link{impulseest}},\code{\link{impulse}}
}

