% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/idframe.R
\name{summary.idframe}
\alias{summary.idframe}
\title{idframe-Object summaries}
\usage{
\method{summary}{idframe}(object)
}
\arguments{
\item{object}{an object of class "idframe"}
}
\description{
Generates a summary of objects inherting from class `idframe`
}

